fix(daemon): fix rtorrent download.finished config
The example configuration in the documentation is incorrect because it
is missing the equals sign on the last argument. This results in
rtorrent never invoking the cross seed script on the download finished
event. It's also difficult to discover, as the error didn't make it to
my log, only appearing in the curses console for a brief second.

Also, I fixed an unclosed parenthesis in the same area.